backend/requirements.txt
Django==2.2.26
graphene-django==2.15.0
validators==0.18.2
django-graphql-jwt==0.3.4
django-cors-headers==3.7.0
requests==2.26.0
gunicorn==20.0.4
psycopg2==2.8.6
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
django-chunked-upload==2.0.0
detect-delimiter==0.1.1
peewee==3.15.1
openpyxl==3.0.10
pandas==1.4.3
xlrd==2.0.1
pika==1.3.0
python-dotenv==0.21.0